chore(ffa): split listener into multiple classes
