+++ src/main/java/dev/revere/alley/game/ffa/listener/FFABlockListener.java
@@ -1,0 +1,42 @@
+package dev.revere.alley.game.ffa.listener;
+
+import dev.revere.alley.Alley;
+import dev.revere.alley.profile.IProfileService;
+import dev.revere.alley.profile.Profile;
+import dev.revere.alley.profile.enums.EnumProfileState;
+import org.bukkit.entity.Player;
+import org.bukkit.event.EventHandler;
+import org.bukkit.event.Listener;
+import org.bukkit.event.block.BlockBreakEvent;
+import org.bukkit.event.block.BlockPlaceEvent;
+
+/**
+ * @author Emmy
+ * @project alley-practice
+ * @since 16/07/2025
+ */
+public class FFABlockListener implements Listener {
+    @EventHandler
+    private void onBlockPlace(BlockPlaceEvent event) {
+        Player player = event.getPlayer();
+        IProfileService profileService = Alley.getInstance().getService(IProfileService.class);
+        Profile profile = profileService.getProfile(player.getUniqueId());
+        if (profile.getState() != EnumProfileState.FFA) {
+            return;
+        }
+
+        event.setCancelled(true);
+    }
+
+    @EventHandler
+    private void onBlockBreak(BlockBreakEvent event) {
+        Player player = event.getPlayer();
+        IProfileService profileService = Alley.getInstance().getService(IProfileService.class);
+        Profile profile = profileService.getProfile(player.getUniqueId());
+        if (profile.getState() != EnumProfileState.FFA) {
+            return;
+        }
+
+        event.setCancelled(true);
+    }
+}
