+++ src/main/java/dev/revere/alley/game/ffa/listener/FFADamageListener.java
@@ -1,0 +1,78 @@
+package dev.revere.alley.game.ffa.listener;
+
+import dev.revere.alley.Alley;
+import dev.revere.alley.base.combat.ICombatService;
+import dev.revere.alley.game.ffa.cuboid.IFFASpawnService;
+import dev.revere.alley.profile.IProfileService;
+import dev.revere.alley.profile.Profile;
+import dev.revere.alley.profile.enums.EnumProfileState;
+import org.bukkit.entity.Player;
+import org.bukkit.entity.Projectile;
+import org.bukkit.event.EventHandler;
+import org.bukkit.event.EventPriority;
+import org.bukkit.event.Listener;
+import org.bukkit.event.entity.EntityDamageByEntityEvent;
+
+/**
+ * @author Emmy
+ * @project alley-practice
+ * @since 16/07/2025
+ */
+public class FFADamageListener implements Listener {
+    /**
+     * Handles the EntityDamageByEntityEvent.
+     * The event is cancelled if the player is in the FFA state and tries to damage another player.
+     *
+     * @param event The EntityDamageByEntityEvent
+     */
+    @EventHandler(priority = EventPriority.MONITOR, ignoreCancelled = true)
+    public void onEntityDamageByEntityMonitor(EntityDamageByEntityEvent event) {
+        if (!(event.getDamager() instanceof Player) || !(event.getEntity() instanceof Player)) return;
+
+        Player player = (Player) event.getEntity();
+        Player attacker = (Player) event.getDamager();
+
+        IProfileService profileService = Alley.getInstance().getService(IProfileService.class);
+        Profile profile = profileService.getProfile(player.getUniqueId());
+        if (profile.getState() != EnumProfileState.FFA) return;
+
+        ICombatService combatService = Alley.getInstance().getService(ICombatService.class);
+        combatService.setLastAttacker(player, attacker);
+    }
+
+    @EventHandler
+    public void onEntityDamageByEntity(EntityDamageByEntityEvent event) {
+        Player attacker;
+
+        if (event.getDamager() instanceof Player) {
+            attacker = (Player) event.getDamager();
+        } else if (event.getDamager() instanceof Projectile) {
+            if (((Projectile) event.getDamager()).getShooter() instanceof Player) {
+                attacker = (Player) ((Projectile) event.getDamager()).getShooter();
+            } else {
+                return;
+            }
+        } else {
+            return;
+        }
+
+        if (!(event.getEntity() instanceof Player)) return;
+        Player victim = (Player) event.getEntity();
+
+        IProfileService profileService = Alley.getInstance().getService(IProfileService.class);
+        Profile profile = profileService.getProfile(victim.getUniqueId());
+        if (profile.getState() != EnumProfileState.FFA) {
+            return;
+        }
+
+        IFFASpawnService ffaSpawnService = Alley.getInstance().getService(IFFASpawnService.class);
+        if ((ffaSpawnService.getCuboid().isIn(victim) && ffaSpawnService.getCuboid().isIn(attacker)) || (!ffaSpawnService.getCuboid().isIn(victim) && ffaSpawnService.getCuboid().isIn(attacker)) || (ffaSpawnService.getCuboid().isIn(victim) && !ffaSpawnService.getCuboid().isIn(attacker))) {
+            ICombatService combatService = Alley.getInstance().getService(ICombatService.class);
+            if (combatService.isPlayerInCombat(victim.getUniqueId()) && combatService.isPlayerInCombat(attacker.getUniqueId())) {
+                return;
+            }
+
+            event.setCancelled(true);
+        }
+    }
+}
